requirements.txt
streamlit
GitPython
google-generativeai
python-dotenv
